Dump the fabric playtest directly to the run directory when it builds
Ensure that the playtest is easier to update on demand by having the output go directly to the dev environment's mods folder. This is still a manual process for the moment that should be automated further, but cuts out the step of having to comment out the dependency for the first run and delete the cached version of it between builds.